Sync sub Travis configuration (#78)
* Stop testing PHP 7.1

This version is EOL.

* Stop testing specific Doctrine versions

The logic was flawed anyway. It would pin ORM and DBAL to the same version, but they have different version ranges. DBAL is now on 2.10 and ORM on 2.7.

* Use `dist: trusty` to fix issue with RabbitMQ

Sometimes, RabbitMQ would not start with error:
```
Failed to start rabbitmq-server.service: Unit rabbitmq-server.service not found.
```

See https://travis-ci.community/t/is-rabbitmq-service-down/4362

* Stop testing Symfony 2.8, 4.0 and 4.1

These versions are EOL.

* Test Symfony 4.4 instead of 4.2

Let's make our life a bit easier and just test:

* Symfony 3.4 LTS
* Symfony 4.4 LTS